refs 125405: make uer-group service independent of ExternalUserGroup
